move TensorShell from core to forward_model; context problem when pipeline loading
