+++ excalibur/cerberus/forward_model.py
@@ -14,6 +14,9 @@ import excalibur
 import excalibur.system.core as syscore
 from excalibur.util.cerberus import crbce, getmmw
 
+import pytensor.graph as tnsrgraph
+import pytensor.tensor as tnsr
+
 # -- GLOBAL CONTEXT FOR PYMC DETERMINISTICS ---------------------------------------------
 from collections import namedtuple
 
@@ -96,6 +99,29 @@ def ctxtupdt(
     )
     # excalibur.cerberus.core.ctxt = excalibur.cerberus.forward_model.ctxt
     return
+
+
+class TensorShell(tnsrgraph.Op):
+    '''
+    GMR: Tensor Shell for custom models
+    Do not touch the name of the methods
+    '''
+
+    def make_node(self, nodes) -> tnsrgraph.Apply:
+        inputs = [tnsr.as_tensor(n) for n in nodes]
+        outputs = [tnsr.vector()]
+        return tnsrgraph.Apply(self, inputs, outputs)
+
+    def perform(
+        self,
+        node: tnsrgraph.Apply,
+        inputs: list[np.ndarray],
+        output_storage: list[list[None]],
+    ) -> None:
+        output_storage[0][0] = np.asarray(LogLikelihood(inputs))
+        return
+
+    pass
 
 
 # GMR: Gregoire s legacy
